Spell Wheel and more.
Options for toggling between the spell wheel and emoji wheel and toggling if hotkey emotes show up.

additional options to change the distance between the spells in the spell wheel and change how far you have to hover forward to select them.

All the settings are saved in a config file.